improve build script

+++ init.js
@@ -1,4 +1,5 @@
 const fs= require("fs");
 const path=require("path")
 const data= fs.readFileSync(path.join(__dirname,"server/sample-db.json"));
-fs.writeFileSync(path.join(__dirname,"server/db.json"),data);
+if ( !fs.existsSync(path.join(__dirname, "server/db.json")))
+	fs.writeFileSync(path.join(__dirname,"server/db.json"),data);
